functions/stripe-webhook/requirements.txt
functions-framework==3.*
requests==2.*
google-auth==2.*
google-api-python-client==2.*
